activity_app/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
flask-login==0.5.0
flask-wtf==0.14.3
wtforms==2.3.3
kafka-python==2.0.2
pandas==1.3.3
jinja2==3.0.3
